cleaned up & added Flask-JWT-Extented package

+++ resources/item.py
@@ -1,6 +1,5 @@
-# import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity, fresh_jwt_required
 from models.item import ItemModel
 
 class Item(Resource):
@@ -8,45 +7,42 @@ class Item(Resource):
     parser.add_argument('price', type=float, required=True, help='This field can not be left blank')
     parser.add_argument('store_id', type=int, required=True, help='every item needs a store id')
         
-    @jwt_required()    # now required to authenticate before route func is executed
+    @jwt_required    # now required to authenticate before route func is executed
     def get(self, name):
         item = ItemModel.find_by_name(name)   #could also be Item.find_by_name
         if item is not None:
             return item.json()
         else:
             return {"message":"item not found"}, 404
-
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return{'message':"An item with the name '{}' already exists.".format(name)}, 400 # request error
         data = Item.parser.parse_args()
-        item = ItemModel(name, data['price'], data['store_id']) # alternative: item = ItemModel(name, **data)
+        item = ItemModel(name, **data)
         try:
             item.save_to_db()
         except:
             return {'message':'An error occurred inserting the item'}, 500  # internal server error 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilage required.'}, 401
+        
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
         return {'message':'item deleted'}
-        """
-        connection = sqlite3.connect('data.db')
-        cursor = connection.cursor()
-        query = "DELETE FROM items WHERE name=?"
-        cursor.execute(query, (name,))
-        connection.commit()
-        connection.close()
-        return {'message':'item deleted'}
-        """
 
     def put(self, name):
         data = Item.parser.parse_args()
         item = ItemModel.find_by_name(name)
         if item is None:
-            item = ItemModel(name, data['price'], data['store_id']) # alternative: item = ItemModel(name, **data)
+            item = ItemModel(name, **data)
         else:
             item.price = data['price']
             item.store_id = data['store_id']
@@ -55,17 +51,10 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [x.json() for x in ItemModel.query.all()]}  # alternative method: return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}
-
-        """
-        connection = sqlite3.connect('data.db')
-        cursor = connection.cursor()
-        query = "SELECT * FROM items"
-        result = cursor.execute(query)
-        items = []
-        for row in result:
-            items.append({'name':row[1],'price':row[2]})
-        connection.close()
-        return {'items':items}
-        """
+        user_id = get_jwt_identity()
+        items = [x.json() for x in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {'items': [x['name'] for x in items], 'message': 'More data available if you log in.'}, 200
